Refactor AccentManager to be an entity system, makes accents ECS.
