+++ Content.Server/Speech/Components/BackwardsAccentComponent.cs
@@ -1,18 +1,10 @@
-﻿using System;
-using Robust.Shared.GameObjects;
+﻿using Robust.Shared.GameObjects;
 
 namespace Content.Server.Speech.Components
 {
     [RegisterComponent]
-    public class BackwardsAccentComponent : Component, IAccentComponent
+    public class BackwardsAccentComponent : Component
     {
         public override string Name => "BackwardsAccent";
-
-        public string Accentuate(string message)
-        {
-            var arr = message.ToCharArray();
-            Array.Reverse(arr);
-            return new string(arr);
-        }
     }
 }
