--- Content.Server/Speech/Components/IAccentComponent.cs
@@ -1,12 +1,0 @@
-﻿namespace Content.Server.Speech.Components
-{
-    internal interface IAccentComponent
-    {
-        /// <summary>
-        /// Transforms a message with the given Accent
-        /// </summary>
-        /// <param name="message">The spoken message</param>
-        /// <returns>The message after the transformation</returns>
-        public string Accentuate(string message);
-    }
-}
